requirements.txt
altair==5.5.0
anyio==4.9.0
arch==7.2.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
filelock==3.18.0
fonttools==4.56.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imbalanced-learn==0.13.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lightgbm==4.6.0
MarkupSafe==3.0.2
matplotlib==3.10.0
narwhals==1.31.0
numpy==2.2.4
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.1.0
plotly==6.0.1
protobuf==5.29.4
pyarrow==19.0.1
pydeck==0.9.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-telegram-bot==22.0
pytz==2025.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.29.0
setuptools==78.1.0
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
statsmodels==0.14.4
streamlit==1.43.2
streamlit-autorefresh==1.0.1
ta==0.11.0
ta-lib==0.6.3
tenacity==9.0.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.4.2
trio==0.29.0
trio-websocket==0.12.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
